Update endpoints conf
